+++ shunting-yard-common/src/main/java/com/hotels/shunting/yard/common/event/AddPartitionEvent.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2018 Expedia Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,8 +24,10 @@ public class AddPartitionEvent extends ListenerEvent {
   private String protocolVersion;
   private String dbName;
   private String tableName;
+  private String tableLocation;
   private Map<String, String> partitionKeys;
   private List<String> partitionValues;
+  private String partitionLocation;
 
   AddPartitionEvent() {}
 
@@ -43,12 +45,20 @@ public class AddPartitionEvent extends ListenerEvent {
     return tableName;
   }
 
+  public String getTableLocation() {
+    return tableLocation;
+  }
+
   public Map<String, String> getPartitionKeys() {
     return partitionKeys;
   }
 
   public List<String> getPartitionValues() {
     return partitionValues;
+  }
+
+  public String getPartitionLocation() {
+    return partitionLocation;
   }
 
 }