+++ shunting-yard-common/src/main/java/com/hotels/shunting/yard/common/io/jackson/ApiarySqsMessageDeserializer.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2018 Expedia Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ package com.hotels.shunting.yard.common.io.jackson;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import com.hotels.shunting.yard.common.ShuntingYardException;
@@ -38,7 +39,10 @@ public class ApiarySqsMessageDeserializer {
   public <T extends ListenerEvent> T unmarshal(String payload) throws ShuntingYardException {
     try {
       log.debug("Unmarshalled payload is: {}", payload);
-      SqsMessage sqsMessage = mapper.readerFor(SqsMessage.class).readValue(payload);
+      SqsMessage sqsMessage = mapper
+          .configure(JsonParser.Feature.ALLOW_UNQUOTED_CONTROL_CHARS, true)
+          .readerFor(SqsMessage.class)
+          .readValue(payload);
       T event = delegateSerDe.unmarshal(sqsMessage.getMessage());
       return event;
     } catch (Exception e) {