+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - TBD
+### Added
+* Support for handling new elements (location and table parameters) coming in from Apiary events.
+* Support for performing metadata only sync in alter table and alter partition event.
+
 ## [0.0.8] - 2019-02-05
 ### Added
 * Support for passing certain configurations like Graphite directly to internal CT instance using `--ct-config` argument.